Fixed Bug with Radar
the assumed position in radians was not correlating to the actual rotation in the calculation of blip position. Fixed by translating into degrees and back into radians which means the program is not confused when quaternion returns 0.5 or -0.5 for the rotation of 60 degrees.
This would happen as 0.5 is a clockwise rotation whereas -0.5 is a counter-clockwise rotation.